lint
SCCS-vsn: sys/kern/sys_generic.c 7.7
SCCS-vsn: sys/ufs/ffs/ffs_vfsops.c 7.11
SCCS-vsn: sys/ufs/ffs/ufs_vfsops.c 7.11
SCCS-vsn: sys/ufs/lfs/lfs_vfsops.c 7.11
SCCS-vsn: sys/ufs/ufs/ufs_vfsops.c 7.11
SCCS-vsn: sys/ufs/ffs/ufs_disksubr.c 7.7
SCCS-vsn: sys/ufs/ufs/ufs_disksubr.c 7.7
SCCS-vsn: sys/sys/malloc.h 7.6

+++ usr/src/sys/ufs/ufs/ufs_disksubr.c
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)ufs_disksubr.c	7.6 (Berkeley) %G%
+ *	@(#)ufs_disksubr.c	7.7 (Berkeley) %G%
  */
 
 #include "param.h"
@@ -187,7 +187,7 @@ setdisklabel(olp, nlp, openmask)
 	if (nlp->d_magic != DISKMAGIC || nlp->d_magic2 != DISKMAGIC ||
 	    dkcksum(nlp) != 0)
 		return (EINVAL);
-	while ((i = ffs(openmask)) != 0) {
+	while ((i = ffs((long)openmask)) != 0) {
 		i--;
 		openmask &= ~(1 << i);
 		if (nlp->d_npartitions <= i)
